Refactor game logic into modular ES architecture
