+++ index.html
@@ -74,6 +74,6 @@
     </div>
   </template>
 
-  <script src="script.js"></script>
+  <script type="module" src="src/main.js"></script>
 </body>
 </html>
