+++ src/core/constants.js
@@ -1,0 +1,5 @@
+export const STORAGE_KEY = 'online-hustle-sim-v1';
+export const MAX_LOG_ENTRIES = 60;
+export const BLOG_CHUNK = 3;
+export const BLOG_INTERVAL_SECONDS = 10;
+export const COFFEE_LIMIT = 3;
