+++ src/game/hustles.js
@@ -1,0 +1,142 @@
+import { createId, formatMoney } from '../core/helpers.js';
+import { addLog } from '../core/log.js';
+import { getHustleState, getState } from '../core/state.js';
+import { addMoney, spendMoney } from './currency.js';
+import { executeAction } from './actions.js';
+import { checkDayEnd } from './lifecycle.js';
+import { spendTime } from './time.js';
+
+export const HUSTLES = [
+  {
+    id: 'freelance',
+    name: 'Freelance Writing',
+    tag: { label: 'Instant', type: 'instant' },
+    description: 'Crank out a quick article for a client. Not Pulitzer material, but it pays.',
+    details: [
+      () => '⏳ Time: <strong>2h</strong>',
+      () => '💵 Payout: <strong>$18</strong>'
+    ],
+    action: {
+      label: 'Write Now',
+      className: 'primary',
+      disabled: () => getState().timeLeft < 2,
+      onClick: () => {
+        executeAction(() => {
+          spendTime(2);
+          addMoney(18, 'You hustled an article for $18. Not Pulitzer material, but it pays the bills!');
+        });
+        checkDayEnd();
+      }
+    }
+  },
+  {
+    id: 'flips',
+    name: 'eBay Flips',
+    tag: { label: 'Delayed', type: 'delayed' },
+    description: 'Hunt for deals, flip them online. Profit arrives fashionably late.',
+    details: [
+      () => '⏳ Time: <strong>4h</strong>',
+      () => '💵 Cost: <strong>$20</strong>',
+      () => '💰 Payout: <strong>$48 after 30s</strong>'
+    ],
+    defaultState: {
+      pending: []
+    },
+    action: {
+      label: 'Start Flip',
+      className: 'primary',
+      disabled: () => {
+        const state = getState();
+        return state.timeLeft < 4 || state.money < 20;
+      },
+      onClick: () => {
+        executeAction(() => {
+          spendTime(4);
+          spendMoney(20);
+          scheduleFlip();
+          addLog('You listed a spicy eBay flip. In 30 seconds it should cha-ching for $48!', 'delayed');
+        });
+        checkDayEnd();
+      }
+    },
+    extraContent: card => {
+      const status = document.createElement('div');
+      status.className = 'pending';
+      status.textContent = 'No flips in progress.';
+      card.appendChild(status);
+      return { status };
+    },
+    update: (_state, ui) => {
+      updateFlipStatus(ui.extra.status);
+    },
+    process: (now, offline) => processFlipPayouts(now, offline)
+  }
+];
+
+export function scheduleFlip() {
+  const flipState = getHustleState('flips');
+  flipState.pending.push({
+    id: createId(),
+    readyAt: Date.now() + 30000,
+    payout: 48
+  });
+}
+
+export function updateFlipStatus(element) {
+  if (!element) return;
+  const flipState = getHustleState('flips');
+  if (!flipState.pending.length) {
+    element.textContent = 'No flips in progress.';
+    return;
+  }
+  const now = Date.now();
+  const nextFlip = flipState.pending.reduce((soonest, flip) =>
+    flip.readyAt < soonest.readyAt ? flip : soonest
+  );
+  const timeRemaining = Math.max(0, Math.round((nextFlip.readyAt - now) / 1000));
+  const label = timeRemaining === 0 ? 'any moment' : `${timeRemaining}s`;
+  const descriptor = flipState.pending.length === 1 ? 'flip' : 'flips';
+  element.textContent = `${flipState.pending.length} ${descriptor} in progress. Next payout in ${label}.`;
+}
+
+export function processFlipPayouts(now = Date.now(), offline = false) {
+  const state = getState();
+  if (!state) return { changed: false };
+  const flipState = getHustleState('flips');
+  if (!flipState.pending.length) {
+    return { changed: false };
+  }
+
+  const remaining = [];
+  let completed = 0;
+  let offlineTotal = 0;
+
+  for (const flip of flipState.pending) {
+    if (flip.readyAt <= now) {
+      completed += 1;
+      if (offline) {
+        state.money += flip.payout;
+        offlineTotal += flip.payout;
+      } else {
+        addMoney(flip.payout, `Your eBay flip sold for $${formatMoney(flip.payout)}! Shipping label time.`, 'delayed');
+      }
+    } else {
+      remaining.push(flip);
+    }
+  }
+
+  flipState.pending = remaining;
+
+  if (!completed) {
+    return { changed: false };
+  }
+
+  const result = { changed: true };
+  if (offline && offlineTotal > 0) {
+    result.offlineLog = {
+      message: `While you were away, ${completed} eBay ${completed === 1 ? 'flip' : 'flips'} paid out. $${formatMoney(offlineTotal)} richer!`,
+      type: 'delayed'
+    };
+  }
+  return result;
+}
