+++ src/game/registry.js
@@ -1,0 +1,9 @@
+import { ASSETS } from './assets.js';
+import { HUSTLES } from './hustles.js';
+import { UPGRADES } from './upgrades.js';
+
+export const registry = {
+  hustles: HUSTLES,
+  assets: ASSETS,
+  upgrades: UPGRADES
+};
