+++ src/ui/elements.js
@@ -1,0 +1,15 @@
+const elements = {
+  money: document.getElementById('money'),
+  time: document.getElementById('time'),
+  timeProgress: document.getElementById('time-progress'),
+  day: document.getElementById('day'),
+  logFeed: document.getElementById('log-feed'),
+  logTemplate: document.getElementById('log-template'),
+  logTip: document.getElementById('log-tip'),
+  hustleGrid: document.getElementById('hustle-grid'),
+  assetGrid: document.getElementById('asset-grid'),
+  upgradeGrid: document.getElementById('upgrade-grid'),
+  endDayButton: document.getElementById('end-day')
+};
+
+export default elements;
